Fix some mainwindow.ui structural issues.
The main problem is that, on resize, the window size and command line
editboxes do not resize proportionaly (specially on shrink). This commit
changes, along with other small things, the UI hierarchy by separating
the main window into a 1x2 (rows x columns) grid, and the left panel is
a 2x1 subgrid (one for Display File management and other for
Navigation).