Provide Nav2D functionality
This is done by letting the arrow buttons change the current renderer
window's position.

While adding this feature, a sutile but very bad error was caught: the
`on_draw` binding was binding a *copy* of the renderer instead of
Rudolph's. This was fixed by using `sigc::mem_fun` instead of a manual
member function binding with `sigc::bind`. Changing `_renderer` to
`&_renderer` in the previous `sigc::bind` call also fixes this issue,
but `sigc::mem_fun` is more meaningful, stating better that line of
code's intention.